ironsight_dashboard/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ironsight-command-center"
version = "0.1.0"
description = "Production-grade real-time rail inspection dashboard integrating 5 neural networks"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Core dependencies
    "numpy>=1.24.0,<2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    # Basic ML dependencies
    "opencv-python-headless>=4.8.0",
    "scikit-learn>=1.3.0",
    "pillow>=10.0.0",
    "einops>=0.7.0",
    # Dashboard dependencies
    "streamlit>=1.28.0",
    "streamlit-image-comparison>=0.0.4",
    "pandas>=2.0.0",
    # Visualization
    "matplotlib>=3.7.0",
    # Performance monitoring
    "psutil>=5.9.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "opencv-python>=4.11.0.86",
    "niquests>=3.16.1",
    "basicsr>=1.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.82.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.uv]
index-strategy = "unsafe-best-match"
# Only resolve dependencies for Windows AMD64
environments = ["sys_platform == 'win32' and platform_machine == 'AMD64'"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
ironsight-dashboard = "src.app:main"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "hypothesis>=6.150.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
